+++ pyaml/bpm/bpm.py
@@ -3,8 +3,8 @@ from typing import Self
 from ..bpm.bpm_model import BPMModel, BPMModelSchema
 from ..common.element import Element, ElementSchema
 from ..common.exception import PyAMLException
-from ..configuration import register_schema
 from ..lattice.abstract_impl import RBpmArray, RWBpmOffsetArray, RWBpmTiltScalar
+from ..validation import register_schema
 
 
 class BPMSchema(ElementSchema):
