+++ pyaml/lattice/simulator.py
@@ -8,7 +8,7 @@ from ..common.abstract_aggregator import ScalarAggregator
 from ..common.element import Element
 from ..common.element_holder import ElementHolder
 from ..common.exception import PyAMLException
-from ..configuration import ConfigurationSchema, get_root_folder, register_schema
+from ..configuration import get_root_folder
 from ..diagnostics.tune_monitor import BetatronTuneMonitor
 from ..lattice.abstract_impl import (
     BPMHScalarAggregator,
@@ -37,6 +37,7 @@ from ..rf.rf_plant import RFPlant, RWTotalVoltage
 from ..rf.rf_transmitter import RFTransmitter
 from ..tuning_tools.measurement_tool import MeasurementTool
 from ..tuning_tools.tuning_tool import TuningTool
+from ..validation import ConfigurationSchema, register_schema
 from .attribute_linker import PyAtAttributeElementsLinker
 from .lattice_elements_linker import LatticeElementsLinker, LatticeElementsLinkerSchema
