+++ pyaml/magnet/hcorrector.py
@@ -1,7 +1,7 @@
 from ..common import abstract
 from ..common.constants import HORIZONTAL_KICK_SIGN
-from ..configuration import register_schema
 from ..lattice.polynom_info import PolynomInfo
+from ..validation import register_schema
 from .corrector import RWCorrectorAngle
 from .magnet import Magnet, MagnetSchema
 from .model import MagnetModel
