+++ pyaml/magnet/linear_cfm_model.py
@@ -3,10 +3,10 @@ from pydantic import BaseModel, ConfigDict
 
 # from ..common.element import __pyaml_repr__
 from ..common.exception import PyAMLException
-from ..configuration import ConfigurationSchema, register_schema
 from ..configuration.curve import Curve, CurveSchema
 from ..configuration.matrix import Matrix, MatrixSchema
 from ..control.deviceaccess import DeviceAccess, DeviceAccessSchema
+from ..validation import ConfigurationSchema, register_schema
 from .model import MagnetModel
