+++ pyaml/magnet/skewquad.py
@@ -1,5 +1,5 @@
-from ..configuration import register_schema
 from ..lattice.polynom_info import PolynomInfo
+from ..validation import register_schema
 from .magnet import Magnet, MagnetSchema
 from .model import MagnetModel
