+++ pyaml/tuning_tools/chromaticity_response_matrix.py
@@ -7,7 +7,7 @@ from pydantic import ConfigDict
 
 from ..common.constants import Action
 from ..common.element import ElementSchema
-from ..configuration import register_schema
+from ..validation import register_schema
 from .measurement_tool import MeasurementTool, MeasurementToolSchema
 from .response_matrix_data import ResponseMatrixDataSchema
