+++ pyaml/tuning_tools/orbit_response_matrix.py
@@ -9,8 +9,8 @@ from pySC.apps.codes import ResponseCode
 
 from ..common.constants import Action
 from ..common.element import ElementSchema
-from ..configuration import register_schema
 from ..external.pySC_interface import pySCInterface
+from ..validation import register_schema
 from .measurement_tool import MeasurementTool, MeasurementToolSchema
 from .orbit_response_matrix_data import OrbitResponseMatrixDataSchema
