+++ pyaml/validation/legacy_registry.json
